reorganize the app entrypoint
